Move process proxy classes to their own package and files.
Moved the processproxy.py file to elyra/services/processproxies and
split out the DistributedProcessProxy and YarnClusterProcessProxy classes
to their own files - distributed.py and yarn.py, respectively.

Updated kernel.json files to reference new location of process proxy class.

Increased default logging level of yarn-api to Warning since each request
was getting logged as INFO.  If troubleshooting is required for the yarn
api, it can be adjusted via ELYRA_YARN_LOG_LEVEL env, e.g., export
ELYRA_YARN_LOG_LEVEL=INFO.
