+++ etc/kernels/spark_2.1_R_yarn_client/kernel.json
@@ -2,7 +2,7 @@
   "language": "R",
   "display_name": "Spark 2.1 - R (YARN Client Mode)",
   "process_proxy": {
-    "class_name": "elyra.services.kernels.processproxy.DistributedProcessProxy"
+    "class_name": "elyra.services.processproxies.distributed.DistributedProcessProxy"
   },
   "env": {
     "SPARK_HOME": "/usr/hdp/current/spark2-client",
